Make PyOBO a core requirement
