+++ setup.cfg
@@ -66,6 +66,7 @@ install_requires =
     tabulate
     pandas
     jsonschema
+    pyobo>=0.2.9
 
 # Random options
 zip_safe = false
@@ -86,8 +87,6 @@ indra =
 jupyter =
     jinja2
     ipython
-grounding =
-    pyobo>=0.2.8
 neo4j =
     py2neo
 docs =
